update localization pt2 pt, ru, zh
- add langauge name
- fix logic to update foreign strings sources
