+++ src/components/SpellLanguagePicker.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
-import { useGlobal } from '../context/GlobalContext';
-import { User } from '../model';
 import {
   List,
   ListItem,
@@ -8,8 +6,7 @@ import {
   ListItemText,
   Checkbox,
 } from '@mui/material';
-import { langName } from '../utils';
-import { useOrbitData } from '../hoc/useOrbitData';
+import { useLocLangName } from '../utils/useLocLangName';
 const ipc = (window as any)?.electron;
 
 interface IProps {
@@ -18,11 +15,9 @@ interface IProps {
 
 export const SpellLanguagePicker = (props: IProps) => {
   const { onSetCodes } = props;
-  const users = useOrbitData<User[]>('user');
-  const [user] = useGlobal('user');
   const [checked, setChecked] = React.useState<string[]>([]);
   const [codes, setCodes] = React.useState<string[]>([]);
-  const [uiLang, setUiLang] = React.useState('en');
+  const [getName, compare] = useLocLangName();
 
   const handleToggle = (value: string) => () => {
     const currentIndex = checked.indexOf(value);
@@ -37,21 +32,6 @@ export const SpellLanguagePicker = (props: IProps) => {
     setChecked(newChecked);
   };
 
-  const getUiName = (value: string, ui: string) => {
-    let res = langName(ui, value) || langName(ui, value.replace('-', '_'));
-    if (res) return res;
-    const items = value.split('-').slice(0, -1);
-    return langName(ui, items.join('_'));
-  };
-
-  const getName = (value: string) => {
-    return getUiName(value, uiLang) || getUiName(value, 'en');
-  };
-
-  const compare = (x: string, y: string) => {
-    return getName(x) < getName(y) ? -1 : 1;
-  };
-
   React.useEffect(() => {
     ipc?.availSpellLangs().then((list: string[]) => {
       setCodes(list);
@@ -61,8 +41,6 @@ export const SpellLanguagePicker = (props: IProps) => {
       setChecked(list);
     });
 
-    const userRec = users.filter((u) => u.id === user) as User[];
-    setUiLang(userRec[0]?.attributes?.locale || 'en');
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
