+++ src/store/localization/exported-strings-name.json
@@ -1,1 +1,1 @@
-{ "stringsName": "strings1e96e6d6.json" }
+{ "stringsName": "stringsd2a52681.json" }
